+++ backend/sh/create_dir.sh
@@ -22,22 +22,19 @@ set -o errexit   # abort on nonzero exitstatus
 set -o pipefail  # don't hide errors within pipes
 
 function usage() {
-    echo "Usage: $0 -i input_file -o output_file -p partition -w wic_image [-g groupname] [-m chmod_mode ] [-u username]" 1>&2; exit 1;
+    echo "Usage: $0 -d dirname -p partition -w wic_image [-g groupname] [-m chmod_mode ] [-u username]" 1>&2; exit 1;
 }
 
-while getopts "i:o:p:w:g:m:u:" opt; do
+while getopts "d:p:w:g:m:u:" opt; do
     case "${opt}" in
-        i)
-            i=${OPTARG}
+        d)
+            d=${OPTARG}
             ;;
         g)
             g=${OPTARG}
             ;;
         m)
             m=${OPTARG}
-            ;;
-        o)
-            o=${OPTARG}
             ;;
         p)
             p=${OPTARG}
@@ -55,20 +52,18 @@ while getopts "i:o:p:w:g:m:u:" opt; do
 done
 shift $((OPTIND-1))
 
-if [ -z "${i}" ] || [ -z "${o}" ] || [ -z "${p}" ] || [ -z "${w}" ]; then
+if [ -z "${d}" ] || [ -z "${p}" ] || [ -z "${w}" ]; then
     usage
 fi
 
-d_echo "i = ${i}"
+d_echo "d = ${d}"
 d_echo "g = ${g}"
 d_echo "m = ${m}"
-d_echo "o = ${o}"
 d_echo "p = ${p}"
 d_echo "u = ${u}"
 d_echo "w = ${w}"
 
 [[ ! -f ${w} ]] && echo "error: input device image not found" 1>&2 && exit 1
-[[ ! -f ${i} ]] && echo "error: input file \"${i}\" not found" 1>&2 && exit 1
 
 # set up loop device to be able to mount image.wic
 losetup_image_wic
@@ -78,10 +73,9 @@ part_pattern="${p}"
 d_echo part_pattern="${p}"
 mount_part
 
-# copy file
-mkdir -p $(dirname /tmp/mount/${p}/${o})
-d_echo "cp ${i} /tmp/mount/${p}/${o}"
-cp ${i} /tmp/mount/${p}/${o}
+# mkdir
+d_echo "mkdir -p /tmp/mount/${p}/${d}"
+mkdir -p /tmp/mount/${p}/${d}
 
 # mount rootA for permission handling
 if [ ! -z "${g}" ] || [ ! -z "${u}" ]; then
@@ -92,25 +86,25 @@ fi
 # group permission handling
 if [ ! -z "${g}" ]; then
     gid=$(cat /tmp/mount/rootA/etc/group | grep ${g}: | awk 'BEGIN { FS = ":" } ; { print $3 }')
-    d_echo $(ls -l /tmp/mount/${p}/${o})
-    d_echo "chgrp ${gid} /tmp/mount/${p}/${o}"
-    chgrp ${gid} /tmp/mount/${p}/${o}
-    d_echo $(ls -l /tmp/mount/${p}/${o})
+    d_echo $(ls -ld /tmp/mount/${p}/${d})
+    d_echo "chgrp ${gid} /tmp/mount/${p}/${d}"
+    chgrp ${gid} /tmp/mount/${p}/${d}
+    d_echo $(ls -ld /tmp/mount/${p}/${d})
 fi
 
 # user permission handling
 if [ ! -z "${u}" ]; then
     uid=$(cat /tmp/mount/rootA/etc/passwd | grep ${u}: | awk 'BEGIN { FS = ":" } ; { print $3 }')
-    d_echo $(ls -l /tmp/mount/${p}/${o})
-    d_echo "chown ${uid} /tmp/mount/${p}/${o}"
-    chown ${uid} /tmp/mount/${p}/${o}
-    d_echo $(ls -l /tmp/mount/${p}/${o})
+    d_echo $(ls -ld /tmp/mount/${p}/${d})
+    d_echo "chown ${uid} /tmp/mount/${p}/${d}"
+    chown ${uid} /tmp/mount/${p}/${d}
+    d_echo $(ls -ld /tmp/mount/${p}/${d})
 fi
 
 # chmod permission handling
 if [ ! -z "${m}" ]; then
-    d_echo $(ls -l /tmp/mount/${p}/${o})
-    d_echo "chmod ${m} /tmp/mount/${p}/${o}"
-    chmod ${m} /tmp/mount/${p}/${o}
-    d_echo $(ls -l /tmp/mount/${p}/${o})
+    d_echo $(ls -ld /tmp/mount/${p}/${d})
+    d_echo "chmod ${m} /tmp/mount/${p}/${d}"
+    chmod ${m} /tmp/mount/${p}/${d}
+    d_echo $(ls -ld /tmp/mount/${p}/${d})
 fi