+++ backend/sh/set_iotedge_gw_config.sh
@@ -3,6 +3,8 @@
 # include shared functions
 . /ics-dm-sh/functions
 
+d_echo ${0}
+
 # exit handler which makes sure we dont leave an undefined host state regarding loop devices
 function finish {
     set +o errexit
@@ -10,6 +12,8 @@ function finish {
     umount /tmp/mount/etc
     umount /tmp/mount/rootA
     losetup -d ${loopdev}
+    while losetup ${loopdev} &>/dev/null; do sleep 0.1; done
+    sync
 }
 trap finish EXIT
 
@@ -20,7 +24,7 @@ function usage() {
 set -o errexit   # abort on nonzero exitstatus
 set -o pipefail  # don't hide errors within pipes
 
-while getopts ":c:e:k:r:w:" opt; do
+while getopts "c:e:k:r:w:" opt; do
     case "${opt}" in
         c)
             c=${OPTARG}
@@ -98,12 +102,8 @@ if [ ! -e /tmp/mount/rootA/etc/systemd/system/multi-user.target.wants/enrollment
     echo "iotedge config apply" >> /tmp/mount/rootA/usr/bin/ics_dm_first_boot.sh
 fi
 
-# set hostname
-hostname=$(grep "^hostname" ${c} | cut -d "=" -f2 | xargs)
-d_echo "set hostname to ${hostname}"
-echo "${hostname}" > /tmp/mount/etc/upper/hostname
-cp /tmp/mount/rootA/etc/hosts /tmp/mount/etc/upper/
-sed -i "s/^127.0.1.1\(.*\)/127.0.1.1 ${hostname}/" /tmp/mount/etc/upper/hosts
+# config hostname
+config_hostname ${c}
 
 # copy root ca cert
 mkdir -p /tmp/mount/data/var/secrets